pyproject.toml
[build-system]
requires = ["setuptools>=61.2"]
build-backend = "setuptools.build_meta"

[project]
name = "kausal-watch"
requires-python = ">= 3.13"
version = "1.0.0"
dependencies = [
    "asgi-cors",
    "babel",
    "celery",
    "channels",
    "channels-redis",
    "coreapi",
    "cryptography",
    "dal-admin-filters",
    "diffsync",
    "django",
    "django-anymail[mailgun,mailjet,sendgrid]",
    "django-autocomplete-light",
    "django-autoslug",
    "django-cors-headers",
    "django-countries",
    "django-environ",
    "django-extensions",
    "django-filter",
    "django-hijack",
    "django-image-cropping",
    "django-modeltrans",
    "django-npm",
    "django-oauth-toolkit",
    "django-parler",
    "django-redis",
    "django-relations-iterator",
    "django-reversion",
    "django-storages[s3]",
    "django-stubs-ext",
    "djangorestframework",
    "dnspython",
    "drf-nested-routers",
    "drf-spectacular",
    "easy-thumbnails",
    "elasticsearch",
    "elasticsearch-dsl",
    "factory-boy",
    "graphene-django",
    "graphene-django-optimizer",
    "graphene-pydantic",
    "graphql-core",
    "graphql-relay",
    "html2text",
    "httpx",
    "humanize",
    "jinja2",
    "libvoikko",
    "logfmter",
    "loguru",
    "lxml",
    "markdown",
    "nanoid",
    "networkx",
    "opentelemetry-exporter-otlp-proto-http",
    "opentelemetry-exporter-prometheus",
    "opentelemetry-instrumentation-django",
    "opentelemetry-instrumentation-psycopg",
    "opentelemetry-instrumentation-redis",
    "opentelemetry-instrumentation-threading",
    "opentelemetry-sdk",
    "orjson",
    "pillow",
    "polars[xlsx2csv]",
    "psutil",
    "psycopg[pool]",
    "pycryptodome",
    "pygithub",
    "pygments-graphql",
    "pytest-cov",
    "pytest-django",
    "pytest-env",
    "pytest-factoryboy",
    "pytest-html",
    "pyyaml",
    "requests",
    "rich",
    "s3cmd",
    "sentry-sdk[opentelemetry]",
    "setuptools<80",
    "social-auth-app-django",
    "starlette",
    "strawberry-graphql",
    "strawberry-graphql-django",
    "tinycss2",
    "treelib",
    "tzdata",
    "unidecode",
    "uritemplate",
    "uuid-utils",
    "uvicorn",
    "wagtail",
    "wagtail-autocomplete",
    "wagtail-color-panel",
    "wagtail-generic-chooser",
    "wagtail-grapple",
    "wagtail-localize",
    "wagtail-modeladmin",
    "wagtail-orderable",
    "wagtailgeowidget",
    "xlsxwriter",
]

[project.optional-dependencies]
kausal = [
    "kausal-watch-extensions>=0.1.156",
]

[tool.ruff]
extend = "./kausal_common/configs/ruff.toml"
extend-exclude = ["./paths_integration/_generated_", "./mcp_server/__generated__"]


[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "aplans.settings"
norecursedirs = [
  ".*",
  ".direnv",
  ".git",
  ".mypy_cache",
  "**/media",
  "**/static",
  "**/templates",
  "Attic",
  "data",
  "node_modules",
  "venv",
]
filterwarnings = ["ignore::DeprecationWarning:easy_thumbnails"]
testpaths = ["**/tests"]
env = [
  "ENV_FILE=.env.test",
]

[tool.mypy]
plugins = [
  "mypy_django_plugin.main",
  "pydantic.mypy",
]
mypy_path = "$MYPY_CONFIG_FILE_DIR/kausal_common/typings"
exclude = [
  "/migrations/",
  "^Attic/",
  "^local_settings\\.py$",
  "^typings/",
  "^kausal_common/deployment/gunicorn\\.py$",
  "node_modules",
  "^.*/_generated_/.*$",
  "^.*/__generated__/.*$",
  #"^kausal_common/typings/",
]
check_untyped_defs = true
cache_fine_grained = true
explicit_package_bases = true
strict_optional = true
fixed_format_cache = true
always_true = ['IS_WATCH']
always_false = ['IS_PATHS']

[[tool.mypy.overrides]]
module = [
  "anymail.*",
  "colored.*",
  "channels.*",
  "dal_admin_filters.*",
  "dal.*",
  "debug_toolbar.*",
  "django_filters.*",
  "django_filters.*",
  "dvc_pandas.*",
  "easy_thumbnails.*",
  "elasticsearch_dsl.*",
  "environ.*",
  "factory.*",
  "generic_chooser.*",
  "graphene_django_optimizer.*",
  "graphene_django.*",
  "grapple.*",
  "gunicorn.*",
  "laces.*",
  "modelcluster.*",
  "modeltrans.*",
  "nanoid.*",
  "networkx.*",
  "oauth2_provider.*",
  "opentelemetry.*",
  "permissionedforms.*",
  "pint_pandas.*",
  "pint.*",
  "plotext.*",
  "rest_framework_nested.*",
  "reversion.*",
  "rustface.*",
  "social_core.*",
  "social_django.*",
  "storages.*",
  "taggit.*",
  "tinycss2.*",
  "traitlets.*",
  "uwsgidecorators.*",
  "viztracer.*",
  "wagtail_localize.*",
  "wagtail_modeladmin.*",
  "wagtail.*",
  "wagtailautocomplete.*",
  "wagtailgeowidget.*",
  "wagtailorderable.*",
  "willow.*",
  "xlsxwriter.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
  "mcp_server.__generated__.*",
  "paths_integration._generated_.*",
]
ignore_errors = true

[tool.mypy-baseline]
baseline_path = ".mypy-baseline.txt"
preserve_position = false
sort_baseline = true

[tool.pyright]
exclude = [
  ".*",
  "**/__pycache__",
  "**/migrations",
  "**/node_modules",
  "Attic/",
  "local_settings.py",
]
stubPath = "kausal_common/typings"
deprecateTypingAliases = true
reportAny = false
reportAttributeAccessIssue = "error"
reportConstantRedefinition = false
reportExplicitAny = false
reportIgnoreCommentWithoutRule = false
reportImplicitOverride = false
reportImplicitStringConcatenation = false
reportImportCycles = false
reportIncompatibleMethodOverride = false
reportIncompatibleVariableOverride = false
reportMissingParameterType = false
reportMissingTypeStubs = "information"
reportPrivateImportUsage = false
reportPrivateUsage = false
reportUnannotatedClassAttribute = false
reportUninitializedInstanceVariable = false
reportUnknownArgumentType = false
reportUnknownMemberType = false
reportUnknownParameterType = false
reportUnknownVariableType = false
reportUnnecessaryComparison = "information"
reportUnnecessaryIsInstance = false
reportUnreachable = false
reportUnsafeMultipleInheritance = "information"
reportUnusedCallResult = false
reportUnusedImport = false
defineConstant = { 'IS_PATHS' = false, 'IS_WATCH' = true }


[tool.django-stubs]
django_settings_module = "aplans.settings"

[tool.setuptools]
include-package-data = false
py-modules = []

[tool.uv]
environments = [
  "python_full_version >= '3.13' and sys_platform == 'linux' and platform_machine == 'x86_64'",
  "python_full_version >= '3.13' and sys_platform == 'darwin' and platform_machine == 'arm64'",
]

[[tool.uv.index]]
url = "https://pypi.org/simple"

[[tool.uv.index]]
name = "kausal"
url = "https://pypi.kausal.tech"
authenticate = "always"

[tool.uv.sources]
graphene-django-optimizer = { url = "https://github.com/kausaltech/graphene-django-optimizer/archive/f18d827d0d58c1148288b6b886c167dd2fc1931a.zip" }
django-modeltrans = { url = "https://github.com/kausaltech/django-modeltrans/archive/a734040c3f38c2e404a61dc0be2c017874a9ea45.zip" }
kausal-paths-extensions = { index = "kausal"}

[dependency-groups]
dev = [
    "ariadne-codegen>=0.16.0",
    "basedpyright>=1.32.1",
    "daphne",
    "django-debug-toolbar",
    "ipdb",
    "ipython",
    "pdbpp",
    "prek>=0.2.12",
    "uv",
    "virtualenvwrapper>=6.1.1",
    "viztracer",
    "watchfiles",
]
kausal = [
    "kausal-watch-extensions",
]
lint = [
    "celery-stubs",
    "django-stubs",
    "djangorestframework-stubs",
    "graphene-stubs",
    "isort",
    "mypy-baseline",
    "mypy[faster-cache]",
    "ondivi",
    "ruff",
    "types-oauthlib",
    "types-psutil",
    "types-python-dateutil",
    "types-uwsgi",
]
mcp = [
    "fastmcp>=2.14.2",
]
prod = [
    "gunicorn",
    "rustface",
    "uvicorn-worker",
    "uvicorn[standard]",
]
